server/backend_requirements.txt
BeautifulSoup==3.2.1
Django==1.7.8
Mako==0.9.1
MarkupSafe==0.18
PAM==0.4.2
Pillow==2.3.0
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
apt-xapian-index==0.45
apturl==0.4.1ubuntu4
argparse==1.2.1
braintree==3.25.0
chardet==2.0.1
colorama==0.2.5
command-not-found==0.3
configglue==1.1.2
debtagshw==0.1
defer==1.0.6
dirspec==13.10
django-adminplus==0.5
django-cors-headers==1.1.0
django-countries==3.4.1
django-crispy-forms==1.6.0
django-tinymce==2.3.0
django-watson==1.2.3
djangorestframework==3.2.3
djoser==0.3.1
dnspython==1.11.1
duplicity==0.6.23
feedparser==5.1.3
html5lib==0.999
httplib2==0.8
lockfile==0.8
lxml==3.3.3
mockito==0.7.0
nemo-emblems==0.0.1
numpy==1.8.2
oauthlib==0.6.1
oneconf==0.3.7
paramiko==1.10.1
pexpect==3.1
piston-mini-client==0.7.5
protobuf==2.5.0
pyOpenSSL==0.13
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pygobject==3.12.0
pyinotify==0.9.4
pyserial==2.6
pysmbc==1.0.14.1
python-apt==0.9.3.5
python-debian==0.1.21-nmu2ubuntu2
pytz==2016.6.1
pyxdg==0.25
reportlab==3.0
requests==2.2.1
sessioninstaller==0.0.0
six==1.5.2
system-service==0.1.6
uTidylib==0.2
urllib3==1.7.1
wsgiref==0.1.2
zope.interface==4.0.5
